app/Executable/parquet2cockroach/requirements.txt
polars
pyarrow
psycopg2-binary
SQLAlchemy
sqlalchemy-cockroachdb
sqlalchemy-utils
PyYAML